Further refactoring for lineage
