+++ packages/components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@labkey/components",
-  "version": "2.172.0-fb-registry-grid-bulk-edit.3",
+  "version": "2.172.0-fb-registry-grid-bulk-edit.5",
   "description": "Components, models, actions, and utility functions for LabKey applications and pages",
   "main": "dist/components.js",
   "module": "dist/components.js",
